% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlmixr_vpc_theme.R
\docType{data}
\name{nlmixr_vpc_theme}
\alias{nlmixr_vpc_theme}
\title{Default VPC theme for 'xpose.nlmixr'}
\format{An object of class \code{vpc_theme} of length 23.}
\usage{
nlmixr_vpc_theme
}
\value{
A list with 'vpc' theme specifiers.
}
\description{
Default VPC theme for 'xpose.nlmixr'.
}
\keyword{datasets}
