% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_nlmixr_model.R
\name{summarise_nlmixr_model}
\alias{summarise_nlmixr_model}
\title{Data summary function}
\usage{
summarise_nlmixr_model(obj, model, software, rounding, runname)
}
\arguments{
\item{obj}{nlmixr fit object to be evaluated}

\item{model}{Model. Can be blank}

\item{software}{Software that generated the model fit}

\item{rounding}{Number of figures to round estimates to}

\item{runname}{Name of the model object being converted}
}
\description{
Convert 'nlmixr' model output into an 'xpose' database
}
\examples{
\dontrun{
xpdb <- xpose_data_nlmixr(obj = fit1)
}
}
