% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_xp_nlmixr.R
\name{theme_xp_nlmixr}
\alias{theme_xp_nlmixr}
\title{Default 'nlmixr' theme for 'xpose'}
\usage{
theme_xp_nlmixr()
}
\value{
A list with 'xpose' theme specifiers.
}
\description{
Default 'nlmixr' theme for 'xpose'.
}
