% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpc_nlmixr.R
\name{vpc_nlmixr}
\alias{vpc_nlmixr}
\title{Create a visual predictive check (VPC) from an 'nlmixr' model}
\usage{
vpc_nlmixr(obj, dat, nsim = 400, by = NULL, ...)
}
\arguments{
\item{obj}{nlmixr fit object to be used as the basis for the VPC.}

\item{dat}{Data file to use as the basis for simulations.}

\item{nsim}{Number of new datasets to simulate (default is 400).}

\item{by}{Variable by which to stratify the VPC. Currently not used.}

\item{...}{Additional arguments to be passed to the \code{\link[vpc]{vpc}} functions.}
}
\description{
Create a visual predictive check (VPC) from an 'nlmixr' model.
}
\examples{
\dontrun{
vpc1 <- vpc_nlmixr(as.saem(myfit), dat, nsim=400, show=list(obs_dv=T),
lloq=0, n_bins = 16, vpc_theme = nlmixr_vpc_theme)
}

}
